backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
bcrypt==4.2.0
qrcode==7.4.2
pillow==11.0.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
